// a3front/src/modules/series-tracking/series-programs/interfaces/form-series-programs.interface.ts
export interface FormSeriesProgramsInterface {
  id?: number | null;
  salidaId: number | null;
  programaId: number | null;
  fileNumber: string;
  passengersCount: number | null;
  groupName: string;
  clienteId: number | null;
  especialistaId: number | null;
  machuPicchuEntry: string;
  observations: string;
  method: string;
}